+++ data/carcMetadata.js
@@ -1,6 +1,6 @@
 export default {
-  version: "1.1.0",
-  releaseDate: "2026-08-29",
+  version: "1.2.0",
+  releaseDate: "2026-08-30",
   source: "Brewer Data Solutions",
   description: "CARC Denial Intelligence Crosswalk",
 };